display a low rank approximation initially
We couldn't find a way to communicate back to the worker until all the computation was done,
which might take a long time. So now, we compute the low rank approximation in the worker separately
from the rest of the computation, and then send the low rank approximation back to the script and display
it. This way, the user can see the low rank approximation immediately, and then waits for the rest